Reorg Serializer into seperate parts.
